# source/packages/contracts/control/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Agent Control API
  version: 0.3.0
  description: |
    Control-plane contract for daemon clients.
    Unknown `/v1/*` routes return `404` problem+json envelopes with correlation headers.
servers:
  - url: http://127.0.0.1:7071
    description: Localhost TCP default binding
security:
  - bearerAuth: []
paths:
...
  /v1/chat/turn:
    post:
      operationId: chatTurn
      summary: Execute one unified orchestrated chat turn.
      parameters:
        - $ref: '#/components/parameters/CorrelationID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatTurnRequest'
      responses:
        '200':
          description: Successful JSON response.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationID'

            X-PersonalAgent-API-Version:
              $ref: '#/components/headers/APIVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatTurnResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '501':
          $ref: '#/components/responses/NotImplemented'
...
  /v1/automation/comm-trigger/metadata:
    post:
      operationId: automationCommTriggerMetadata
      summary: Return typed ON_COMM_EVENT trigger metadata/schema contracts.
      parameters:
        - $ref: '#/components/parameters/CorrelationID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutomationCommTriggerMetadataRequest'
      responses:
        '200':
          description: Successful JSON response.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationID'

            X-PersonalAgent-API-Version:
              $ref: '#/components/headers/APIVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationCommTriggerMetadataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/automation/comm-trigger/validate:
    post:
      operationId: automationCommTriggerValidate
      summary: Validate and normalize ON_COMM_EVENT filter authoring payloads.
      parameters:
        - $ref: '#/components/parameters/CorrelationID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutomationCommTriggerValidateRequest'
      responses:
        '200':
          description: Successful JSON response.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationID'

            X-PersonalAgent-API-Version:
              $ref: '#/components/headers/APIVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationCommTriggerValidateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '501':
          $ref: '#/components/responses/NotImplemented'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  parameters:
    CorrelationID:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
      description: Optional correlation ID propagated across daemon responses/events.
...
  headers:
    CorrelationID:
      description: Correlation ID associated with this response.
      schema:
        type: string
    APIVersion:
      description: Daemon API contract version served by this endpoint.
      schema:
        type: string
        example: v1
...
  responses:
...
    BadRequest:
      description: Invalid request payload.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationID'
        X-PersonalAgent-API-Version:
          $ref: '#/components/headers/APIVersion'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid bearer token.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationID'
        X-PersonalAgent-API-Version:
          $ref: '#/components/headers/APIVersion'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotImplemented:
      description: |
        Service route is registered but backing domain service is not configured.
        `error.code` is `service_not_configured` when applicable and `error.details`
        includes typed `service`, `domain`, and `remediation` metadata.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationID'
        X-PersonalAgent-API-Version:
          $ref: '#/components/headers/APIVersion'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    AutomationCommTriggerMetadataRequest:
      type: object
      properties:
        workspace_id:
          type: string
    AutomationCommTriggerRequiredDefaults:
      type: object
      properties:
        event_type:
          type: string
        direction:
          type: string
        assistant_emitted:
          type: boolean
    AutomationCommTriggerKeywordFilter:
      type: object
      properties:
        contains_any:
          type: array
          items:
            type: string
        contains_all:
          type: array
          items:
            type: string
        exact_phrases:
          type: array
          items:
            type: string
    AutomationCommTriggerFilter:
      type: object
      properties:
        channels:
          type: array
          items:
            type: string
        principal_actor_ids:
          type: array
          items:
            type: string
        sender_allowlist:
          type: array
          items:
            type: string
        thread_ids:
          type: array
          items:
            type: string
        keywords:
          $ref: '#/components/schemas/AutomationCommTriggerKeywordFilter'
    AutomationCommTriggerFilterFieldSchema:
      type: object
      properties:
        field:
          type: string
        value_type:
          type: string
        match_semantics:
          type: string
        description:
          type: string
    AutomationCommTriggerMetadataCompatibility:
      type: object
      properties:
        principal_filter_behavior:
          type: string
        keyword_match_behavior:
          type: string
    AutomationCommTriggerMetadataResponse:
      type: object
      properties:
        trigger_type:
          type: string
        required_defaults:
          $ref: '#/components/schemas/AutomationCommTriggerRequiredDefaults'
        idempotency_key_fields:
          type: array
          items:
            type: string
        filter_defaults:
          $ref: '#/components/schemas/AutomationCommTriggerFilter'
        filter_schema:
          type: array
          items:
            $ref: '#/components/schemas/AutomationCommTriggerFilterFieldSchema'
        compatibility:
          $ref: '#/components/schemas/AutomationCommTriggerMetadataCompatibility'
    AutomationCommTriggerValidateRequest:
      type: object
      properties:
        workspace_id:
          type: string
        subject_actor_id:
          type: string
        filter:
          $ref: '#/components/schemas/AutomationCommTriggerFilter'
    AutomationCommTriggerValidationIssue:
      type: object
      properties:
        code:
          type: string
        field:
          type: string
        message:
          type: string
    AutomationCommTriggerValidationCompatibility:
      type: object
      properties:
        compatible:
          type: boolean
        subject_actor_id:
          type: string
        subject_matches_principal_rule:
          type: boolean
    AutomationCommTriggerValidateResponse:
      type: object
      properties:
        valid:
          type: boolean
        trigger_type:
          type: string
        required_defaults:
          $ref: '#/components/schemas/AutomationCommTriggerRequiredDefaults'
        normalized_filter:
          $ref: '#/components/schemas/AutomationCommTriggerFilter'
        normalized_filter_json:
          type: string
        errors:
          type: array
          items:
            $ref: '#/components/schemas/AutomationCommTriggerValidationIssue'
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/AutomationCommTriggerValidationIssue'
        compatibility:
          $ref: '#/components/schemas/AutomationCommTriggerValidationCompatibility'
...
    ChatTurnRequest:
      type: object
      properties:
        workspace_id:
          type: string
        task_class:
          type: string
        requested_by_actor_id:
          type: string
        subject_actor_id:
          type: string
        acting_as_actor_id:
          type: string
        provider:
          type: string
        model:
          type: string
        system_prompt:
          type: string
        channel:
          $ref: '#/components/schemas/ChatTurnChannelContext'
        tool_catalog:
          type: array
          items:
            $ref: '#/components/schemas/ChatTurnToolCatalogEntry'
        items:
          type: array
          items:
            $ref: '#/components/schemas/ChatTurnItem'
    ChatTurnResponse:
      type: object
      properties:
        workspace_id:
          type: string
        task_class:
          type: string
        provider:
          type: string
        model_key:
          type: string
        correlation_id:
          type: string
        contract_version:
          type: string
        turn_item_schema_version:
          type: string
        realtime_event_contract_version:
          type: string
        channel:
          $ref: '#/components/schemas/ChatTurnChannelContext'
        items:
          type: array
          items:
            $ref: '#/components/schemas/ChatTurnItem'
        task_run_correlation:
          $ref: '#/components/schemas/ChatTurnTaskRunCorrelation'
...
    ChatTurnToolCatalogEntry:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        capability_keys:
          type: array
          items:
            type: string
        input_schema:
          type: object
          additionalProperties: true
...
    ChatTurnChannelContext:
      type: object
      properties:
        channel_id:
          type: string
          description: Canonical logical channel id (`app`, `message`, or `voice`).
        connector_id:
          type: string
          description: Optional connector attribution for the turn context.
        thread_id:
          type: string
          description: Optional channel-thread correlation id.
    ChatTurnItem:
      type: object
      properties:
        item_id:
          type: string
        type:
          type: string
          description: Canonical turn item type (`user_message`, `assistant_message`, `tool_call`, `tool_result`, `approval_request`, `approval_decision`).
        role:
          type: string
        status:
          type: string
        content:
          type: string
        tool_name:
          type: string
        tool_call_id:
          type: string
        arguments:
          type: object
          additionalProperties: true
        output:
          type: object
          additionalProperties: true
        error_code:
          type: string
        error_message:
          type: string
        approval_request_id:
          type: string
        metadata:
          type: object
          additionalProperties: true
    ChatTurnTaskRunCorrelation:
      type: object
      properties:
        available:
          type: boolean
        source:
          type: string
        task_id:
          type: string
        run_id:
          type: string
        task_state:
          type: string
        run_state:
          type: string
...
    ErrorResponse:
      type: object
      additionalProperties: true
      required:
        - error
        - correlation_id
        - type
        - title
        - status
        - detail
        - instance
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              description: |
                Optional domain-specific typed error metadata.
                For `service_not_configured` errors this includes `category`,
                `service` (`id`, `label`, `config_field`), `domain`, and `remediation`
                (`action`, `label`, `hint`).
        correlation_id:
          type: string
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string